spring-api-starter code

+++ src/main/java/com/codewithmosh/store/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.codewithmosh.store.repositories;
 import com.codewithmosh.store.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Byte> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 }
